removing index from labelquant

+++ lib/fil/fil_test.go
@@ -49,8 +49,8 @@ func Test_readPepXMLInput(t *testing.T) {
 				t.Errorf("readPepXMLInput() got1 = %v, want %v", got1, tt.want1)
 			}
 
-			if got[0].Index != uint32(18992) {
-				t.Errorf("Index is incorrect, got %d, want %d", got[0].Index, uint32(18992))
+			if got[0].Index != 18992 {
+				t.Errorf("Index is incorrect, got %d, want %d", got[0].Index, 18992)
 			}
 
 			if got[0].Spectrum != "b1906_293T_proteinID_01A_QE3_122212.60782.60782.2#interact.pep.xml" {
